Currency.tsx #13
This is a rough draft of what the virtual coin system is, it is similar to gambling apps like fliff which does not use real world currency but app credits to bet.
This takes on issue #13, seems fine for now but will need to be wired to purchase a bet once that has been created.